+++ TESTING/LIN/clattp.c
@@ -36,7 +36,7 @@ static real c_b93 = 2.f;
     void c_div(complex *, complex *, complex *);
     double pow_dd(doublereal *, doublereal *), sqrt(doublereal);
     void r_cnjg(complex *, complex *);
-    double c_abs(complex *);
+    float c_abs(complex *);
 
     /* Local variables */
     real c__;
